Moved `PluginLoader` to its own file
