+++ src/lib.rs
@@ -318,6 +318,7 @@ mod context;
 mod engine;
 mod main_loop;
 mod plugin;
+mod plugin_loader;
 mod scheduler;
 mod state;
 mod state_stack;
@@ -336,6 +337,7 @@ pub use context::*;
 pub use engine::*;
 pub use main_loop::*;
 pub use plugin::*;
+pub use plugin_loader::*;
 pub use scheduler::*;
 pub use state::*;
 pub use state_stack::*;
